Add support for service termination protection
Termination protection can be useful for preventing accidental deletion
of production services. Because it only prevents deletion of the service
itself and not any sub resources one must still be extremely careful
with the diffs when applying changes but at least it will ensure that
backups are not lost for services that do keep backups.