Do not wait for Kafka auxiliary services to start up
The check was already skipped in many cases because IP filter or VPC
may prevent connecting to the aux services to check their state.
Because the check is often skipped nothing can really rely on it so
the value it adds is very limited. The Aiven API should be extended
to provide more detailed state information for the aux services to
make this work properly.

This may possibly fix https://github.com/aiven/terraform-provider-aiven/issues/95
although I was unable to reproduce that.